+++ pycep/transformer.py
@@ -996,6 +996,17 @@ class BicepToJson(Transformer[Token, pycep_typing.BicepJson]):
             },
         }
 
+    def ends_with(self, args: tuple[pycep_typing.PossibleValue, pycep_typing.PossibleValue]) -> pycep_typing.EndsWith:
+        string_to_search, string_to_find = args
+
+        return {
+            "type": "ends_with",
+            "parameters": {
+                "string_to_search": string_to_search,
+                "string_to_find": string_to_find,
+            },
+        }
+
     def format(self, args: tuple[pycep_typing.PossibleValue, ...]) -> pycep_typing.Format:
         format_string, *arg_x = args